[fix] - lotto 수정
