+++ src/main/java/LottoSimulatorApplication.java
@@ -1,6 +1,5 @@
 import LottoSimulatorHandler.LottoSimulatorHandler;
 import java.io.IOException;
-import service.LottoSimulatorService;
 
 public class LottoSimulatorApplication {
   public static void main(String[] args) throws IOException {
