+++ src/main/java/model/Number.java
@@ -8,5 +8,4 @@ public class Number {
 
   public static final int ONE =1;
   public static final int ZERO =0;
-
 }
